client: Make more types public
