+++ crates/wasm-pkg-client/src/loader.rs
@@ -1,5 +1,3 @@
-use std::cmp::Ordering;
-
 use async_trait::async_trait;
 use bytes::Bytes;
 use futures_util::{stream::BoxStream, StreamExt};
@@ -8,37 +6,7 @@ use wasm_pkg_common::{
     Error,
 };
 
-use crate::Release;
-
-#[derive(Clone, Debug, Eq)]
-pub struct VersionInfo {
-    pub version: Version,
-    pub yanked: bool,
-}
-
-impl Ord for VersionInfo {
-    fn cmp(&self, other: &Self) -> Ordering {
-        self.version.cmp(&other.version)
-    }
-}
-
-impl PartialOrd for VersionInfo {
-    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
-        Some(self.version.cmp(&other.version))
-    }
-}
-
-impl PartialEq for VersionInfo {
-    fn eq(&self, other: &Self) -> bool {
-        self.version == other.version
-    }
-}
-
-impl std::fmt::Display for VersionInfo {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{version}", version = self.version)
-    }
-}
+use crate::release::{Release, VersionInfo};
 
 #[async_trait]
 pub trait PackageLoader: Send {
